MachineKnight - House Rent Predictor/App/requirements.txt
Pillow==9.2.0
streamlit==1.11.0
sklearn<=1.1.2s